3_Source_Code/requirements.txt
AWSIoTPythonSDK==1.4.9
pyserial==3.5
python-dotenv==0.19.0
boto3==1.34.59

